add javadocs
