+++ src/main/java/com/google/graphgeckos/dashboard/fetchers/github/GitHubController.java
@@ -8,12 +8,12 @@ public class GitHubController {
   /**
    * Handles POST requests from the GitHub API. Setting GitHub Webhooks is required.
    * Extracts required data from the json {@see CommitData} and adds it to the Storage.
-   * @param commitData json from the GitHub API.
+   * @param gitHubData json from the GitHub API.
    */
   @RequestMapping(value = "/github-info", method = RequestMethod.POST, headers = {"content-type=application/json"})
-  public CommitData postGitHubInfo(@RequestBody CommitData commitData) {
+  public GitHubData postGitHubInfo(@RequestBody GitHubData gitHubData) {
     // TODO(issue #64): implement described functionality.
-    return commitData;
+    return gitHubData;
   }
 
 }
