+++ src/test/resources/jsons/output_no_unknown_fields_json.txt
